requirements.txt
pandas
numpy
scikit-learn
flask
flask_restful
scipy
sqlalchemy
PyYAML